# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "api for making short links",
        "title": "Link shortener API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/api/get/{short}": {
            "get": {
                "description": "Get original link by short equivalent",
                "produces": [
                    "application/json"
                ],
                "summary": "Get original link by short equivalent",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Short",
                        "name": "short",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.GetResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/api/make": {
            "post": {
                "description": "Make short link",
                "produces": [
                    "application/json"
                ],
                "summary": "Make short link",
                "parameters": [
                    {
                        "description": "Make request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.MakeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.MakeResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        }
    },
    "definitions": {
        "controllers.GetResponse": {
            "type": "object",
            "properties": {
                "original": {
                    "type": "string",
                    "example": "http://example.com/"
                }
            }
        },
        "controllers.MakeRequest": {
            "type": "object",
            "required": [
                "original",
                "short"
            ],
            "properties": {
                "original": {
                    "type": "string",
                    "maxLength": 1000,
                    "minLength": 1,
                    "example": "http://example.com/"
                },
                "short": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 6,
                    "example": "da3rsf"
                }
            }
        },
        "controllers.MakeResponse": {
            "type": "object",
            "properties": {
                "short": {
                    "type": "string",
                    "example": "da3rsf"
                }
            }
        }
    }
}